+++ ECSJobDemos/Assets/ECS/Spawners/SpawnRandomCircleSystem.cs
@@ -25,7 +25,7 @@ namespace ECS.Spawners
             var maingroup = EntityManager.CreateComponentGroup(typeof(SpawnRandomCircle),typeof(TransformPosition));
             maingroup.CompleteDependency();
 
-            EntityManager.GetAllUniqueSharedComponents(uniqueTypes);
+            EntityManager.GetAllUniqueSharedComponentDatas(uniqueTypes);
 
             int spawnInstanceCount = 0;
             for (int sharedIndex = 0; sharedIndex != uniqueTypes.Count; sharedIndex++)
@@ -93,10 +93,10 @@ namespace ECS.Spawners
                     {
                         position = positions[i]
                     };
-                    EntityManager.SetComponent(entities[i],position);
+                    EntityManager.SetComponentData(entities[i],position);
                 }
 
-                EntityManager.RemoveSharedComponent<SpawnRandomCircle>(sourceEntity);
+                EntityManager.RemoveComponent<SpawnRandomCircle>(sourceEntity);
                 
                 entities.Dispose();
                 positions.Dispose();